feat(sepolia): set the Jovian parameters on Base Sepolia (#515)
This task sets the minimum base fee to 200,000 wei (2e-4 gwei) and the DA footprint scalar to 312. These are the values we intend to use on Base Mainnet as well.

200,000 wei was chosen based on the p5 base fee observed on Base Mainnet from November 7 to 14 and rounded down slightly to get a nice number. The DA footprint was calculated from `da_footprint_gas_scalar = block_gas_target_throughput / (blob_target * 128,000 bytes / 12 seconds * estimation_ratio)` where `block_gas_target_throughput = 30,000,000 gas/sec`, `blob_target = 6`, and `estimation_ratio = 1.5`. `30,000,000 / (6 * 128,000 / 12 * 1.5) = 312.5`, which was rounded down to the nearest integer.

The task uses the `OWNER_SAFE` hierarchy from the `sepolia/2025-11-07-upgrade-fault-proofs` task. This differs from SystemConfig parameter changes on Base Mainnet, which do not use the nested safes.